Fix PriceSnapshot model issues
- Add shop uniqueness constraints to support multiple shops
- Set price precision to 4 decimal places instead of 2
- Make serializer handle duplicate insertions gracefully by using get_or_create

+++ server/products/utils.py
@@ -1,10 +1,11 @@
 from datetime import datetime
+from decimal import Decimal
 
 def format_date(date_string):
     return datetime.strptime(date_string, '%m/%d/%Y')
 
 def convert_price_to_float(price_string):
-    return float(price_string.replace('$', ''))
+    return Decimal(price_string.replace('$', ''))
 
 def validate_unit(unit):
     if unit not in ['lb', 'oz', 'gal', 'ea', 'g', 'kg', 'ml', 'l']:
